+++ CHANGELOG.md
@@ -10,5 +10,6 @@ All notable changes to this project are documented here. The format is based on
 -   Added ISO 8601 -> DateTime pipe.
 -   Added RFC 2822 -> DateTime pipe.
 -   Added arbitrary format -> DateTime pipe.
+-   Added DateTime -> ISO 8601 pipes (full/date/time/week).
 
 [keep-a-changelog]: http://keepachangelog.com/en/1.0.0/